Dyn::Call::Aggregate and Dyn::Call::Field
Dyn::Call::Pointer->perl( ... ) will need the reverse at some point

+++ t/src/42_simple_args.c
@@ -10,7 +10,7 @@ DLLEXPORT int TakeTwoShorts(short x, short y) {
     return 0;
 }
 
-DLLEXPORT int AssortedIntArgs(int x, short y, char z) {
+DLLEXPORT long AssortedIntArgs(long x, short y, char z) {
     if (x == 101 && y == 102 && z == 103) return 3;
     return 0;
 }
